newcardtemplate props
